requirements.txt
flask
openai
feedreader
psycopg2
feedparser
azure-cognitiveservices-speech
tiktoken
requests
../python-client
